Add fetched MOTD for TFC
This adds a bit to the Login command, fetching a dynamic message of the
day on successful login if the user is logging in to Terraform Cloud or
a slightly more custom experience for Terraform Enterprise.

It's purposely very fail-open - if the request for a login message fails
in any way at all, we log the error as a trace and just use a vanilla
welcome message. MOTD should never be an error-raising issue if it
doesn't work.